[td] Fix sample datasets
